Add files via upload
This is the final capstone project completed via IBM Data Science from Coursera. 
It is a compilation of neighborhood and crime data from San Francisco outlining the best
area for a potential client to open a new restaurant! 
I apologize in advance if it seems confusing, wrong, or if some of the code is redundant. 
I'm still learning! :)